functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp==3.8.5
asyncio==3.4.3
azure-functions==1.15.0
azure-storage-blob==12.17.0
msal==1.23.0
